requirements.txt
functions-framework==3.*
google-cloud-bigquery
google-cloud-storage
pymongo==4.3.3
pandas==1.5.3
requests==2.28.1
beautifulsoup4==4.12.2
